Compare process hashes based on original process name
We need to compare process hashes based on the original
task name rather than on a name that may have been altered
if it was a kernel thread.

Add in a new field "task_mangled" which is a modified version
of the task name if it is a kernel thread. This field is the
field that is printed, where as task is the field that is used
internally for task hashing.

Signed-off-by: Colin Ian King <colin.king@canonical.com>