Improve kernel thread detection and hash on unmodified task name (LP: #1467932)
The kernel thread detection now checks in an empty command line first
then pgid being zero and if all else false falls back to a table lookup
on possible kernel threads. This works better for containerised and non
containerised enviroments.

Also, keep two copies of the task name, the original, which we hash on
and the mangled name (which gets [ ] bracketed if process is a kernel
thread).

Finally, pass a timer_info_t struct to timer_stat_add rather than
a lot of arguments as this reduces the register/stack count on this
heavily used function and makes it a bit cleaner too.

Signed-off-by: Colin Ian King <colin.king@canonical.com>