Add some performance optimisations, better hashing and heap management
Use some GCC optimisations to make sure the hottest code paths are
optimised higher.  Don't keep on free'ing and allocating timer stats
from the heap but keep a cache hot free list so we can keep on re-using
the same memory.   Reduce the hash table size as traversing a large
empty hash to free all the items is causing a lot of cache stalls.

We now have the top 67% of the code now in kernel space fetching the
timer stats raw data with these optimisations, so we're saving a lot
of CPU for some of the lower power devices we run this tool on.

Signed-off-by: Colin Ian King <colin.king@canonical.com>