Add auto-purge of old state timer stats to stop memory growth
eventstat was not originally designed to run for long periods
and so it accumulated timer stats.  The downside to this is that
the timer stat traveral did not scale since this was a linear list
and also it grew overtime.

This change adds two required features. The first converts the
timer stat lookup into a hashed lookup, so this speeds up this
even for many timer stats.  The second feature is a auto-purging
of timer stats that have not been referenced for 600 seconds. This
should ensure that keep the cache small.

Signed-off-by: Colin Ian King <colin.king@canonical.com>